requirements.txt
adversarial-robustness-toolbox==1.9.1
altgraph==0.17.2
certifi==2021.10.8
colorama==0.4.4
ffmpeg-python==0.2.0
future==0.18.2
joblib==1.1.0
kornia==0.6.2
llvmlite==0.38.0
numba==0.55.1
numpy==1.21.5
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.0
pefile==2021.9.3
Pillow==9.0.0
pyinstaller==4.8
pyinstaller-hooks-contrib==2022.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
pywin32-ctypes==0.2.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.0.0
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
typing_extensions==4.0.1
wincertstore==0.2
